Strip bank_id's last char
CSV import create duplicate entries because it relies and the `bank_id`
field to always be the same for a transaction, across exports.

Observations show a transaction always has the same `bank_id` across
exports, **except** for the last char. I haven't found any documentation
about the CSV/bank_id format but looking a the duplicate entries we have
in the database and at entries from a couple of difference CSV exports,
it should fix our issue.